Exact single stage Phase II design

Calculates the sample size and decision rule for exact single stage Phase II design

Usage

ph2single(pu,pa,ep1,ep2,nsoln=5)

Arguments

pu

unacceptable response rate; baseline response rate that needs to be exceeded for treatment to be deemed promising

pa

response rate that is desirable; should be larger than pu

ep1

threshold for the probability of declaring drug promising under pu (target type 1 error rate); between 0 and 1

ep2

threshold for the probability of declaring the drug not promising under pa (target type 2 error rate); between 0 and 1

nsoln

number of designs to be returned for given ep1 and ep2

Value

ph2single returns a data frame with the following variables:

n - the total number of subjects treated (n)

r - the number of responses (r) needed to be exceeded to consider treatment promising

Type I error

Type II error

Details

The function assesses multiple designs and returns the nsoln possible ones with the type-I and type-II errors closest to and below the target ones given in arguments (ep1, ep2). The treatment is deemed promising if > r responses are observed in the n patients.

See also

Examples

ph2single(0.2, 0.4, 0.1, 0.1)
#>    n  r Type I error Type II error
#> 1 36 10   0.08891278    0.09036317
#> 2 39 11   0.07422615    0.08821755
#> 3 40 11   0.08750524    0.07094949
#> 4 42 12   0.06210708    0.08595484
#> 5 43 12   0.07332134    0.06954899
ph2single(0.2, 0.35, 0.05, 0.05)
#>     n  r Type I error Type II error
#> 1  98 26   0.04458319    0.04705221
#> 2  99 26   0.04997882    0.04072296
#> 3 101 27   0.03848785    0.04857528
#> 4 102 27   0.04322253    0.04215896
#> 5 103 27   0.04837386    0.03648695
ph2single(0.2, 0.35, 0.05, 0.05, nsoln = 10)
#>      n  r Type I error Type II error
#> 1   98 26   0.04458319    0.04705221
#> 2   99 26   0.04997882    0.04072296
#> 3  101 27   0.03848785    0.04857528
#> 4  102 27   0.04322253    0.04215896
#> 5  103 27   0.04837386    0.03648695
#> 6  105 28   0.03736318    0.04355357
#> 7  106 28   0.04188931    0.03779612
#> 8  107 28   0.04681002    0.03271037
#> 9  108 29   0.03228543    0.04490739
#> 10 109 29   0.03625671    0.03907135