Simon's 2-stage Phase II design

Calculates Optimal and Minimax 2-stage Phase II designs given by Richard Simon

ph2simon(pu, pa, ep1, ep2, nmax=100)
# S3 method for ph2simon
print(x, ...)
# S3 method for ph2simon
plot(x, ...)

Arguments

pu
unacceptable response rate

pa
response rate that is desirable

ep1
threshold for the probability of declaring drug desirable under p0

ep2
threshold for the probability of rejecting the drug under p1

nmax
maximum total sample size (default 100; can be at most 500)

x
object returned by ph2simon

...
arguments to be passed onto plot and print commands called within

Value

ph2simon returns a list with pu, pa, alpha, beta and nmax as above and:

out

matrix of best 2 stage designs for each value of total sample size n. the 6 columns are: r1, n1, r, n, EN(p0), PET(p0)

Trial is stopped early if <= r1 responses are seen in the first stage and treatment is considered desirable only when >r responses seen. The "print" method formats and returns the minimax and optimal designs. The "plot" plots the expected sample size agains the maximum sample size as in Jung et al., 2001
See also

References

Simon R. (1989). Optimal Two-Stage Designs for Phase II Clinical Trials. Controlled Clinical Trials 10, 1-10.

Jung SH, Carey M and Kim KM. (2001). Graphical Search for Two-Stage Designs for Phase II Clinical Trials. Controlled Clinical Trials 22, 367-372.

Examples

  ph2simon(0.2, 0.4, 0.1, 0.1)

#> 
#>  Simon 2-stage Phase II design 
#> 
#> Unacceptable response rate:  0.2 
#> Desirable response rate:  0.4 
#> Error rates: alpha =  0.1 ; beta =  0.1 
#> 
#>         r1 n1  r  n EN(p0) PET(p0)
#> Optimal  3 17 10 37  26.02  0.5489
#> Minimax  3 19 10 36  28.26  0.4551
#> 

  ph2simon(0.2, 0.35, 0.05, 0.05)

#> 
#>  Simon 2-stage Phase II design 
#> 
#> Unacceptable response rate:  0.2 
#> Desirable response rate:  0.35 
#> Error rates: alpha =  0.05 ; beta =  0.05 
#> 
#>         r1 n1  r  n EN(p0) PET(p0)
#> Optimal 11 53 26 99  69.88  0.6330
#> Minimax 15 68 25 95  75.44  0.7244
#> 

#> Warning:   Optimal sample size too close to nmax. 
#>   Try increasing nmax (current value = 100)

  ph2simon(0.2, 0.35, 0.05, 0.05, nmax=150)

#> 
#>  Simon 2-stage Phase II design 
#> 
#> Unacceptable response rate:  0.2 
#> Desirable response rate:  0.35 
#> Error rates: alpha =  0.05 ; beta =  0.05 
#> 
#>         r1 n1  r   n EN(p0) PET(p0)
#> Optimal 11 50 28 109  67.07  0.7107
#> Minimax 15 68 25  95  75.44  0.7244
#>